backend/requirements.txt
Flask
pymongo
psycopg2-binary
redis
pika
celery
